Renamed NameFilter files to JobNameFilter name structure
