--- build-monitor-plugin/src/main/java/com/smartcodeltd/jenkinsci/plugins/buildmonitor/viewmodel/features/nameFilter/NameFilterConfig.java
@@ -1,14 +1,0 @@
-package com.smartcodeltd.jenkinsci.plugins.buildmonitor.viewmodel.features.nameFilter;
-
-/**
- * @author Vincent & Robert
- */
-public class NameFilterConfig {
-    public final String prefix, suffix, regex;
-
-    public NameFilterConfig(String prefix, String suffix, String regex) {
-        this.prefix = prefix;
-        this.suffix = suffix;
-        this.regex = regex;
-    }
-}
